[#39] Add Zuul job for GitHub Issues integration
Adds a post job to update related issues in Github on change merges. If
a change contains a "Closes: #X" tag, issue X will be closed on merge.
If a change contains a "Relates-To: #X" or "[#X]" tag, a comment will be
made on the associated issue to indicate the change was merged and that
the issue may be closeable.

Closes: #39

Change-Id: Ide9a4d53de7be0e467422938e4e8067fb41ce036