[#58] Add golint
This adds the golint linter to the `make lint` target. As of this
commit, the results from golint are not counted as errors. Further
changes to the code will be required to bring it up to golint's
standards, at which point we will allow golint to cause failures for the
`lint` target.

Change-Id: I34134e3deb080e6da0bc0aac299caca6c3b5d0a9