Fix read the docs post job
The previous setup here was rejected by zuul as an invalid
configureation because trigger-readthedocs-webhook is defined as a final
job. Final jobs cannot be modified or parented (which is what the old
config was doing). Unfortunately, this prevented Zuul from running any
post jobs as it has to read the configuration for all jobs in a pipeline
as a whole.

We can fix this by feeding the trigger-readthedocs-webhook job the
necessary vars to run from project level vars. This doesn't modify the
job itself and only serves to add inputs to it.

Change-Id: Ibc3a62db4277aaf31956783aaef52be6f7a247b5