Add the dynamic kubernetes client
The dynamic client will be needed to interact with any custom resource
that airshipctl doesn't know about. It will be required for checking the
health of a cluster, as well as any other operations that may need to be
performed on generic objects.

This also adds the pkg/k8s/client/fake package, which can be used to
create a mock instance of a client for use in unit tests.

Change-Id: Ia331ff4875a067045f6f9245daee109126fb1d33
Relates-To: #73
Relates-To: #20