Remove the various Equals methods from pkg/config
This change removes the `Equals` methods from the types defined in
pkg/config. They are not being used anywhere in the code base and as
such are dead code.

A user who wishes to compare objects from pkg/config should defer to
using reflect.DeepEqual (several of the objects in pkg/config were
already using DeepEqual, ending up with the same result).

Change-Id: I12bb88b4bf1646f597af70fc8cca93e8d9f7c653