Fix Redfish error handling logic
A logic error in the ScreenRedfishError function causes a false success
if there is no error unmarshalling the body of a GenericOpenAPIError.
This change corrects the logic error and modifies some status checks to
use the http library constants for readability.

This functionality remains untested. The body field of the
GenericOpenAPI error is unexported; therefore, it is inaccessible to our
unit tests.

Fixes: #142

Change-Id: Ic9f248c239830e0b6013a2e2c5d74ea4d2ef25af
Signed-off-by: Drew Walters <andrew.walters@att.com>