Add validation to Redfish media ejection
Currently, airshipctl relies on a wait period after ejecting media
before trying to insert new media; however, this may not be safe if the
sleep time is not long enough.

This change replaces the sleep time with validation logic that polls the
media status before attempting to insert new media.

Change-Id: I64e8892686dbfa2fc780039331c341e14cc6752c
Signed-off-by: Drew Walters <andrew.walters@att.com>