Add system power status check to remotedirect
Invoking remotedirect on a host that is powered off will result in a
remotedirect failure; therefore, the power status of a host needs to be
verified before remotedirect can begin. This change adds a status check
of the ephemeral node before performing remotedirect. When the node is
off, airshipctl powers the node on.

NOTE: if the host does not power on quickly enough, remotedirect could
still fail. The redfish client will be updated to handle the startup and
shutdown waiting in another change.

Change-Id: I725cf79e070864956e2991f60261328ebd6a5fc5
Signed-off-by: Drew Walters <andrew.walters@att.com>