Fix power command host selectors
Both host selectors are always passed to create a manager when a power
command is invoked; however, an empty label selector will include all
hosts.

This change adds a helper function to determine which selectors will be
used based on the flags passed to airshipctl. When a flag is not used, a
selector is not created for it.

Change-Id: I48684cdf6d16073bf85468f6286a22a0aba602f7
Signed-off-by: Drew Walters <andrew.walters@att.com>