Add mutual exclusivity to manager host selectors
Using a label selector and name when creating a manager will produce the
union of all host documents that match the criteria. While this is not
inherently a problem, it's not expected behavior. This change makes
using the label selector and name property mutually exclusive. When
creating a manger using both selectors, the manager will contain the
intersection of hosts, rather than the union.

Change-Id: I791e0460cea2ae3d60ea2cc24f900b925c36d9af
Signed-off-by: Drew Walters <andrew.walters@att.com>