Add state validation to Redfish shutdown/startup
The Redfish client reports success when Redfish returns success after a
shutdown or startup signal is sent; however, airshipctl needs to verify
that the host successfully reaches the desired power state before moving
on. This change adds the same wait functionality from reboot to the
power on and power off functionality.

Change-Id: I56c6696394b547f97da79cec56726002c3f225db
Signed-off-by: Drew Walters <andrew.walters@att.com>