Fixed local scripts
Added remote_work_dir for template to use the same config options
that we have for regular gate.

Without this change when we run local srcipts we have a conflict
when 'airshipctl cluster init' tried to use manifests from '/tmp/airship'
which is targetPath from airconfig. But 'airshipctl document pull'
downloded manifests to '/tmp/airship/airshipctl'

Change-Id: I1eaa2f7525471c756f79a0bc1dd566124f40d531