Add hardware profile / firmware plumbing
This adds firmware definition into the hostgenerator-m3 function,
along with a default hardwareProfile/firmware definition;
it also adds a hardwareprofile-example function that can be used
as an example for creating and consuming hardwareProfiles.

In the future, other things will be added to hardwareProfiles in
addition to firmware, for consumption by hostgenerator-m3
and/or by other functions.

Closes: #291

Change-Id: I586284f07e5b98353af06ecb7559a64771297e44