Re-enable doc validation gate
This sets the validation gate to be voting again; it had been turned
off while some work was in flight.  However, the check is passing again
and re-enabling it will help ensure document goodness.

A couple of tweaks were made based on issues seen during multi-cluster
testing:

1. The kind cluster is now torn down between each cluster (eph and target).
   We'd been reusing the cluster for the sake of speed, but deleting
   everything out of it still left some unexplained cruft behind,
   which broke validation of the second (target) cluster.

2. This reverts back to a simple approach to phase selection and sequencing,
   away from a fancier technique which was failing the only phase
   was "initinfra".

Note: once `airshipctl phase validate` and a queryable phase map are
in place, the gate should be reworked to leverage them instead.

Change-Id: I37c51e859e2df880981e2e138a67c23e17eabd93