Gate: avoid failing before Nodes are defined
This is a workaround for the kubernetes issue below, which has
caused gate failures a few times.  The issue is due to a race condition
between a new apiserver becoming live, and its resource definitions being
fully loaded, resulting in a "no resource type node" kind of error
when running `kubectl wait node`.

This refactors the "loop until apiserver is ready" logic to treat
"no resource found" errors the same as apiserver unavailability,
resulting in another round of `sleep`.

https://github.com/kubernetes/kubernetes/issues/83242

Change-Id: I9b1aa0c0c12bbc9399e5a1f22390074319151df3