Use proxy settings from env vars
For the labs behind the corporate proxy we need to define proxy settings as env vars.
Inside Ironic pod there is init-bootstrap container which is using curl with external
URLs. This change introduces the env vars for proxy in the ConfigMap mounted to all
Ironic containers in the pod. Also Ephemeral Secret now has an option to use proxy.

Relates-To: #355

Change-Id: I4f1c61c8381d57e4ad8adc063434468fdfa0e959