Fix incorrect links in Hardware Profile Documentation
There was an issue that pointed out incorrect references to the
metal3 API within the Hardware Profile section of Airship 2.0
documentation. They are currently pointing to the API docs and
not the API itself. This patch addresses that issue and updates
the link as suggested.

I've kept the last reference intact, since it points to the API
documentation itself and doesn't require a change.

Relates-To: #422
Closes: #422
Change-Id: I37254a64a27b3cdf2675721b0f1f62d41913915c