Helm Chart Collator - Airship Helm Repository
This patchset provides the implementation for the Helm Chart Collator,
which will be used as Helm Repository containing the helm charts for
Airship workloads, e.g., Dex.

This patchset requires the presence of the Docker image
'quay.io/airshipit/helm-chart-collator:latest' which is built by the
project 'airshipit/images' and supported by 'helm-chart-collator'.

The image contains 'baked' charts such as 'dex-aio' charts. If you need
a different set of charts, you will have to create a dedicated docker
image and have it deployed by the Helm Chart Collator service.

Change-Id: Ie62f306ea1c619468d0d45120b23cce7bfda2efd